% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfrappe.R
\name{frappeOutput}
\alias{frappeOutput}
\title{frappeOutput}
\usage{
frappeOutput(outputId, width = "100\%", height = "250px")
}
\arguments{
\item{outputId}{The name of the input}

\item{width}{Must be a valid CSS unit (e.g. "400px", "auto") or a number, which will be coerced to a string and have "px" appended.}

\item{height}{Must be a valid CSS unit (e.g. "400px", "auto") or a number, which will be coerced to a string and have "px" appended.}
}
\description{
Widget output function for use in Shiny
}
