% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfrappe.R
\name{renderFrappe}
\alias{renderFrappe}
\title{renderFrappe}
\usage{
renderFrappe(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An rfrappe expression.}

\item{env}{An enviorment.}

\item{quoted}{A boolean value.}
}
\description{
Widget render function for use in Shiny
}
