% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfrappe.R
\name{rfrappe}
\alias{rfrappe}
\title{rfrappe}
\usage{
rfrappe(options, width = NULL, height = NULL)
}
\arguments{
\item{options}{list of chart options}

\item{width}{does nothing right now}

\item{height}{does nothing right now}
}
\description{
create a Frappe chart widget
}
\examples{
rfrappe(list(
  data = list(
    labels = c("12am-3am", "3am-6pm", "6am-9am", "9am-12am", "12pm-3pm", "3pm-6pm", "6pm-9pm", "9am-12am"),
    datasets = list(
      list(
        title = "Some Data",
        color = "light-blue",
        values = c(25, 40, 30, 35, 8, 52, 17, -4)
      ),
      list(
        title = "Another Set",
        color = "violet",
        values = c(25, 50, -10, 15, 18, 32, 27, 14)
      )
    )
  )
))

}
